+++ docs/CHANGELOG.md
@@ -37,6 +37,7 @@ Version 1.7.0 (not released yet)
 - Add `mute_all_players` and `mute_player` commands for hiding chat messages from specific players
 - Add server config option to give max ammo to the player when a weapon item is picked up
 - Append dedicated server config name to the log file name so logs from multiple servers won't mix
+- Change log entry time format
 
 Version 1.6.1 (released 2021-02-20)
 -----------------------------------